Current std plans are <500 turns
The random message for >500 turns on d1 references a 1d attempt. But
all current plans for 1d standard right now are for 450-496 turns.
Therefore this can be changed to >400 so that it will be seen. Of
course this is only going to be seen by a tiny collection of people,
but it is still funny.