Exclude any blocked monster when generating potential faxes
This intends to teach SFaxGeneratePotentialFaxes to generate faxes by
$monster so that each addition can be evaluated against the
blocked_monster argument.

This argument is needed for things like Genie wishes which only
support a subset of wishes.

In doing this, I found it convenient to provide some structure for
each potential fax. Unfortunately, this doesn't handle all of the
special variations yet like lists or the combined white lion /
whitesnake targets.