Remove banish code duplication and fix ice house
BanishesActive() does exactly the same, and there is no need for the code to be repeated. Same with BanishTurnsLeft, which can be inserted inside DescribeThisBanish.

Ice House did not appear on the banishers list ever because its length was -1 and it was not ending on rollover. While the fixes above also patched it up, added a special string to show that it is really forever.