Fix combat rate calculations
I got confused because it is named "combat rate", so needing to increase exactly by "combat rate" seems weird. But it actually was a "noncombat rate" all along. Replaced it with canonical combat_percent.

There is a bug with Smut Orc Logging Camp (stuck on 0% combat sometimes), but this will need fixing in Mafia.